requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.2
billiard==4.2.2
cachetools==6.2.0
celery==5.5.3
certifi==2025.8.3
channels==4.3.1
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.184.0
google-auth==2.41.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.10
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
stripe==13.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
daphne
channels-redis
redis
psycopg2-binary   
whitenoise